used authorization, and implemented on, delete vieo , and delete user only person who is the owner can delete the video/user, else not

+++ express/backend1/controllers/deleteUser.js
@@ -2,11 +2,17 @@ const { User } = require("../models/User");
 
 const deleteUser = async (req, res, next) => {
   try {
+    if(req.params.id!=req.user.id){
+      return res.status(403).json({
+        message:"fobidden", 
+        success:false
+      })
+    }
     const deletedUser = await User.findByIdAndDelete(req.params.id);
     if (!deletedUser) {
       return res.status(404).json({
         message: "user not found",
-        success: true,
+        success: false,
       });
     }
     res.status(200).json({
